move components to dir
