+++ website/src/components/Schema/index.js~HEAD
@@ -1,10 +1,13 @@
 import React, { useState, useMemo, useEffect } from 'react';
 import Table from '@icgc-argo/uikit/Table';
-import TagButton, { TAG_TYPES } from './TagButton';
+import Tag, { TAG_TYPES } from '../Tag';
 import styles from './styles.module.css';
-import Tag from '@icgc-argo/uikit/Tag';
+import DefaultTag from '@icgc-argo/uikit/Tag';
 import CodeList from './CodeList';
 import Regex from './Regex';
+import startCase from 'lodash/startCase';
+import Button from '@icgc-argo/uikit/Button';
+import Typography from '@icgc-argo/uikit/Typography';
 
 const formatFieldType = value => {
   switch (value) {
@@ -15,6 +18,15 @@ const formatFieldType = value => {
   }
 };
 
+const HeaderName = ({ name }) => {
+  const sentenceCase = startCase(name);
+  return (
+    <h2 className={styles.schemaTitle}>
+      {sentenceCase} ({name})
+    </h2>
+  );
+};
+
 const FieldDescription = ({ name, description }) => (
   <div className={styles.fieldDescription}>
     <div className={styles.name}>{name}</div>
@@ -23,14 +35,15 @@ const FieldDescription = ({ name, description }) => (
 );
 
 const FieldsTag = ({ fieldCount }) => (
-  <Tag className={styles.fieldsTag}>{`${fieldCount} Field${fieldCount > 1 ? 's' : ''}`}</Tag>
+  <DefaultTag className={styles.fieldsTag}>{`${fieldCount} Field${
+    fieldCount > 1 ? 's' : ''
+  }`}</DefaultTag>
 );
 
-const Schema = ({ schema, key, ...props }) => {
+const Schema = ({ schema, key }) => {
   // SSR fix
   if (typeof schema === 'undefined') return null;
 
-  // console.log('schema', typeof schema, schema.name, schema.fields.length);
   /**
    * need to pass in state for Cell rendering
    * react-table rerenders everything, change shape of codelist to pass in state
@@ -59,16 +72,39 @@ const Schema = ({ schema, key, ...props }) => {
     {
       Header: 'Field & Description',
       id: 'fieldDescription',
-      accessor: ({ name, description }) => (
+      Cell: ({ original: { name, description } }) => (
         <FieldDescription name={name} description={description} />
       ),
     },
-    { Header: 'Data Tier' },
+    {
+      Header: 'Data Tier',
+      Cell: ({ original: { meta } }) => {
+        if (meta && meta.primaryId) {
+          return <Tag type={TAG_TYPES.id} />;
+        } else if (meta && meta.core) {
+          return <Tag type={TAG_TYPES.core} />;
+        } else {
+          return <Tag type={TAG_TYPES.extended} />;
+        }
+      },
+    },
     {
       Header: 'Attributes',
       id: 'attributes',
-      accessor: ({ restrictions }) =>
-        restrictions && restrictions.required && <TagButton type={TAG_TYPES.required} />,
+      Cell: ({ original: { restrictions, meta } }) => (
+        <div>
+          {restrictions && restrictions.required && (
+            <div>
+              <Tag type={TAG_TYPES.required} />
+            </div>
+          )}
+          {meta && !!meta.dependsOn && (
+            <div style={{ marginTop: '2px' }}>
+              <Tag type={TAG_TYPES.dependency} />
+            </div>
+          )}
+        </div>
+      ),
     },
     { Header: 'Type', id: 'valueType', accessor: ({ valueType }) => formatFieldType(valueType) },
     {
@@ -76,10 +112,11 @@ const Schema = ({ schema, key, ...props }) => {
       id: 'permissibleValues',
       accessor: 'restrictions',
       Cell: ({ original }) => {
-        const { name: field, restrictions = {} } = original;
+        const { name: field, restrictions = {}, meta } = original;
         const { regex = null, codeList = null } = restrictions;
+        const examples = meta && meta.examples && meta.examples.split(',');
         if (regex) {
-          return <Regex regex={regex} />;
+          return <Regex regex={regex} examples={examples} />;
         } else if (codeList) {
           return (
             <CodeList
@@ -93,19 +130,34 @@ const Schema = ({ schema, key, ...props }) => {
         }
       },
     },
-    { Header: 'Notes & Scripts' },
+    {
+      Header: 'Notes & Scripts',
+      Cell: ({ original: { meta, restrictions } }) => {
+        const script = restrictions && restrictions.script;
+        return (
+          <div>
+            {meta && meta.notes && <div>{meta.notes}</div>}
+            {script && <Button>View Script</Button>}
+          </div>
+        );
+      },
+    },
   ];
   const containerRef = React.createRef();
-  const prefix = 'prefix_prefix';
-  const ext = 'tsv';
 
   return (
     <div className={styles.schema}>
-      <h2 className={styles.schemaTitle}>{schema.name}</h2>
+      <HeaderName name={schema.name} />
       <FieldsTag fieldCount={schema.fields.length} />
-      <div className={styles.fieldExample}>
-        Field Name Example: <span>{`${prefix}`}</span>[-optional-extension]<span>{`.${ext}`}</span>
+      <div>
+        <Typography variant="data">
+          {schema.description}
+          <div className={styles.fieldExample}>
+            Field Name Example: <span>{`${schema.name}`}</span>[-optional-extension].tsv
+          </div>
+        </Typography>
       </div>
+
       <div ref={containerRef}>
         <Table
           parentRef={containerRef}
