--- website/src/pages/dictionary/Bar/styles.module.css
@@ -1,29 +1,0 @@
-.bar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 8px 15px;
-  background: var(--argo-grey-4);
-  border: solid 1px var(--argo-grey-2);
-  margin-bottom: 30px;
-}
-
-.bar .dataSelectors > div {
-  margin: 0 7px 0 5px;
-}
-
-.bar [role='button'],
-.search > div {
-  min-height: 28px;
-  height: 28px;
-}
-
-.bar .search svg {
-  height: 16px;
-  width: 16px;
-}
-
-.bar > div {
-  display: flex;
-  align-items: center;
-}
